First pass at populating graph from Currency ArrayList.
